Convert useIsApplePayAvailable to TypeScript and move to file
